+++ src/main/java/com/kobe/warehouse/service/dto/records/StoreInventoryLineRecord.java
@@ -1,13 +1,11 @@
 package com.kobe.warehouse.service.dto.records;
 
-import java.math.BigInteger;
-
 public record StoreInventoryLineRecord(
     int produitId,
     String produitCip,
     String produitEan,
     String produitLibelle,
-    BigInteger id,
+    Long id,
     Integer gap,
     Integer quantityOnHand,
     int quantityInit,
